Split app.js phase 4c-3: extract blocking-platform module
App-blocking warning overlay, Android/iOS blocking init, friction gate,
warning coordinator, handset modal screens, platform detection.
In-app tests pass (132/132).
