+++ src/enforcement.js
@@ -28,10 +28,9 @@ import { isModalVisible } from './modal-manager.js';
 import { kickClockNow } from './render.js';
 import { isScheduleSegmentActiveNow } from './schedule-editor.js';
 import {
-    applySafariFdaOnboardingLanguage,
-    reconcileBlockingWarningShell, setLanguagePickerOpen, showExclusiveOnboardingScreen,
-    updateOnboardingVisibility,
+    applySafariFdaOnboardingLanguage, setLanguagePickerOpen,
 } from './app.js';
+import { reconcileBlockingWarningShell, showExclusiveOnboardingScreen, updateOnboardingVisibility } from './blocking-platform.js';
 import {
     EXT_ONBOARDING_DISMISSED_KEY, MIGRATION_POLL_MS, applyEnforcementDescCopy,
     hasAcceptedEula, showMigrationOnboarding, presentWelcomeOnboarding,
