+++ src/schedule-editor.js
@@ -11,8 +11,8 @@ import { disableScheduleControls, disableTimeControls, pad, parseEndTimeBoundedI
 import { syncSchedulePanelOverlayControls } from './schedule-overlay.js';
 import {
     shouldUseCompactMobileScheduleDayLabels,
-    updateWindowHeight,
 } from './app.js';
+import { updateWindowHeight } from './blocking-platform.js';
 import { openScheduleOverrideModal, setBtnActionLabel, setStartBlockBtnLeadingIcon, setStartBtnBlocklistInfo, showScheduleConfirmModal, showScheduleEditConfirmModal, syncStopBtnLabelFit, updatePauseButtonAppearance } from './confirm-modals.js';
 
 export const TIME_SEPARATOR_ARROW_HTML = '<span class="time-separator" aria-hidden="true"><svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2.5" stroke-linecap="round" stroke-linejoin="round"><path d="M5 12h14"></path><path d="M13 6l6 6-6 6"></path></svg></span>';
